Build Script & PML Info File
Add `VERSION` to repo root, with the required PML version.

Add `build.sh` script to convert the PML Ref Man after checking that
(1) PML is installed on the user machine and (2) that it matches the
required version stored in the `VERSION` file, issuing a warning if
any of these conditions are not met.